+++ Snowflake.Data/Client/SnowflakeDbDataReader.cs
@@ -262,9 +262,20 @@ namespace Snowflake.Data.Client
                 {
                     throw new StructuredTypesReadingException($"Method GetObject<{typeof(T)}> can be used only for structured object");
                 }
-                var stringValue = GetString(ordinal);
-                var json = stringValue == null ? null : JObject.Parse(stringValue);
-                return JsonToStructuredTypeConverter.ConvertObject<T>(fields, json);
+                var val = GetValue(ordinal);
+                switch (val)
+                {
+                    case string stringValue:
+                        {
+                            var json = JObject.Parse(stringValue);
+                            return JsonToStructuredTypeConverter.ConvertObject<T>(fields, json);
+                        }
+                    case Dictionary<string, object> structArray:
+                        return ArrowConverter.ConvertObject<T>(structArray);
+                    default:
+                        logger.Debug($"Unexpected value of type {val?.GetType().FullName} at ordinal {ordinal}.");
+                        return null;
+                }
             }
             catch (Exception e)
             {
@@ -286,10 +297,20 @@ namespace Snowflake.Data.Client
                 {
                     throw new StructuredTypesReadingException($"Method GetArray<{typeof(T)}> can be used only for structured array or vector types");
                 }
-
-                var stringValue = GetString(ordinal);
-                var json = stringValue == null ? null : JArray.Parse(stringValue);
-                return JsonToStructuredTypeConverter.ConvertArray<T>(fields, json);
+                var val = GetValue(ordinal);
+                switch (val)
+                {
+                    case string stringValue:
+                        {
+                            var json = stringValue == null ? null : JArray.Parse(stringValue);
+                            return JsonToStructuredTypeConverter.ConvertArray<T>(fields, json);
+                        }
+                    case List<object> listArray:
+                        return ArrowConverter.ConvertArray<T>(listArray);
+                    default:
+                        logger.Debug($"Unexpected value of type {val?.GetType().FullName} at ordinal {ordinal}.");
+                        return null;
+                }
             }
             catch (Exception e)
             {
@@ -309,10 +330,20 @@ namespace Snowflake.Data.Client
                 {
                     throw new StructuredTypesReadingException($"Method GetMap<{typeof(TKey)}, {typeof(TValue)}> can be used only for structured map");
                 }
-
-                var stringValue = GetString(ordinal);
-                var json = stringValue == null ? null : JObject.Parse(stringValue);
-                return JsonToStructuredTypeConverter.ConvertMap<TKey, TValue>(fields, json);
+                var val = GetValue(ordinal);
+                switch (val)
+                {
+                    case string stringValue:
+                        {
+                            var json = stringValue == null ? null : JObject.Parse(stringValue);
+                            return JsonToStructuredTypeConverter.ConvertMap<TKey, TValue>(fields, json);
+                        }
+                    case Dictionary<object, object> mapArray:
+                        return ArrowConverter.ConvertMap<TKey, TValue>(mapArray);
+                    default:
+                        logger.Debug($"Unexpected value of type {val?.GetType().FullName} at ordinal {ordinal}.");
+                        return null;
+                }
             }
             catch (Exception e)
             {