+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - Users can now specify non-string values in Toml. For example, `port` can be specified as an integer in the Toml.
     - Add retry for HTTP 307/308 status codes
     - Added exception handling to session heartbeat to prevent network errors from disrupting background heartbeat check.
+    - Added support for native arrow structured types.
 - v5.1.0
     - Added `APPLICATION_PATH` to `CLIENT_ENVIRONMENT` sent during authentication to identify the application connecting to Snowflake.
     - Renew idle sessions in the pool if keep alive is enabled.